test(requests): extract shared token_for / post_action helpers (#40)
## Summary

token_for and the JSON post_action were copy-pasted across the request suite
(token_for in 7 files, post_action byte-identical in 6). Extract them into one
shared support module so a change to token signing edits one file, not seven.

- New spec/requests/support/action_request_helpers.rb defines token_for +
  post_action (the JSON-body POST).
- rails_helper.rb auto-loads requests/support/**/*.rb (mirroring system_helper's
  system/support autoload) and includes the module for type: :request.
- The 6 fully-covered specs drop their local defs.
- post_multipart stays LOCAL to the multipart specs (it sends FormData, not the
  shared JSON body) and reuses the included token_for.

The shared post_action signature (klass, act:, payload: {}, params: {}) and
token_for(klass, payload = {}) satisfy every call site unchanged — all callers
pass keyword args, so the keyword reorder and relaxed-arity are behavior-neutral.

## Test Coverage

- No behavior change — all request specs continue to pass against the shared
  helpers (verified: spec/requests green).

## Verification

- [x] bundle exec rspec spec/phlex spec/requests — 203 examples, 0 failures
- [x] bundle exec standardrb — clean

Refs #34, PR #38. Closes #40
